#178 - More flexible backend creation ABM cluster and NSM clusters can be provided when creating an ALBA Backend Added unit tests for this new scenario Moved unit test helper code to more logic location
