+++ ovs/extensions/storage/volatile/dummystore.py
@@ -38,7 +38,7 @@ class DummyVolatileStore(object):
         """
         self._keep_in_memory_only = True
 
-    def clean(self):
+    def _clean(self):
         """
         Empties the store
         """
